[refactor] Remove unnecessary package
